pyproject.toml
[tool.poetry]
name = "tsumiki"
version = "0.1.0"
description = ""
authors = ["None"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
pip = "^23.3.2"
install = "^1.3.5"
frozen-flask = "^1.0.1"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
ruff = "^0.5.4"
mypy = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
